fixes for rubocop and feedback
